添加用户登陆、注册功能，并完成测试

+++ src/main/java/cn/edu/hebut/ego/entity/Users.java
@@ -46,7 +46,7 @@ public class Users extends Model<Users> {
     private Integer status;
 
     @ApiModelProperty(value = "会员积分")
-    private Long memberPoint;
+    private Integer memberPoint;
 
     @ApiModelProperty(value = "收件人姓名")
     private String receiveName;
